// src/app/api/service/personal-finance-api.service.ts
export interface CategorieDto {
    id?: number;
    name?: string | undefined;
    isShared?: boolean;
    userId?: number;
}

export enum TransactionTypeEnum {
    Einnahmen = "Einnahmen",
    Ausgaben = "Ausgaben",
}

export interface TransactionDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    value?: number;
    date?: Date;
    userId?: number;
    categoryId?: number;
    transactionType?: TransactionTypeEnum;
}

export interface DateRangeDto {
    startDate?: Date;
    endDate?: Date;
}